# inteli-tractian-project/docs/api-contract.openapi.yaml
openapi: 3.1.0
info:
  title: Tractian Industrial Support API (Challenge Inteli)
  description: |
    API industrial simplificada para o Challenge "Engenharia e Avaliação de Agentes Industriais".
    Dados sintéticos, anonimizados. Serve como contexto comum às duas partes do projeto
    (construção e avaliação do agente).

    ## Comportamento probabilístico
    Os endpoints de **consulta** (GET) podem retornar variações controladas por `seed`:
    - `complete` — retorno completo
    - `partial` — informação parcial (campos/etapas faltantes)
    - `inconclusive` — resultado inconclusivo
    - `conflict` — conflito entre fontes
    - `unavailable` — indisponibilidade temporária

    O `seed` (query param) torna o comportamento determinístico para reprodutibilidade na
    avaliação. Sem `seed`, o comportamento é amostrado por uma distribuição fixa.

    ## Ações de alto impacto
    As operações de escrita (POST/PATCH de impacto) exigem **parâmetros válidos + `justification`**
    (string ≥ 20 chars justificando a ação). Uma chamada aceita representa a execução e retorna
    sucesso, **sem ciclo adicional de status**.

    ## Autenticação
    `x-user-id` header identifica a pessoa usuária (perfil/permissões) do contexto. Sem token.
  version: 0.1.0
  contact:
    name: TRACTIAN x Inteli — Challenge
servers:
  - url: http://localhost:8000
    description: Ambiente local de demonstração

# ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    UserContext:
      type: apiKey
      in: header
      name: x-user-id
      description: ID da pessoa usuária (define perfil/permissões e contexto de empresa).
  parameters:
    Seed:
      name: seed
      in: query
      description: Semente para tornar o comportamento probabilístico determinístico.
      required: false
      schema: { type: string }
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string, examples: [usr_001] }
        name: { type: string, examples: [Ana Mantovani] }
        role: { type: string, enum: [operator, mechanic, reliability_analyst, maintenance_manager, coordinator, engineer, electrician] }
        permissions:
          type: array
          items: { type: string, enum: [read, action_low, action_high, escalate] }
        company_id: { type: string }
    AssetConfig:
      type: object
      description: Configuração técnica do ativo (usada em contexto de diagnóstico).
      properties:
        machine_type: { type: string, examples: [motor_induction, motor_dc, pump, compressor, gearbox, fan, spindle, conveyor] }
        rotation_rpm: { type: number, examples: [1780] }
        bearing_specs:
          type: object
          properties:
            part_number: { type: string, examples: [NU 310] }
            bpfo_hz: { type: number }
            bpfi_hz: { type: number }
            bsf_hz: { type: number }
            ftf_hz: { type: number }
        line_frequency_hz: { type: number, examples: [60] }
...
    QueryEnvelope:
      type: object
      description: Envelope de consulta com modo probabilístico.
      properties:
        mode: { type: string, enum: [complete, partial, inconclusive, conflict, unavailable] }
        notes: { type: string, nullable: true }
        data: {} # payload específico do endpoint
    ActionRequest:
      type: object
      required: [justification]
      properties:
        justification:
          type: string
          minLength: 20
          description: Justificativa da ação (≥ 20 caracteres).
        params:
          type: object
          description: Parâmetros específicos da ação.
          additionalProperties: true
    ActionResult:
      type: object
      properties:
        accepted: { type: boolean }
        action_id: { type: string }
        message: { type: string }
paths:
...
  /companies/{companyId}:
    get:
      tags: [Contexto]
      operationId: getCompany
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: companyId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Empresa.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryEnvelope' }
        '404': { description: Não encontrado. }
  /companies/{companyId}/assets:
    get:
      tags: [Contexto]
      operationId: listAssetsByCompany
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: companyId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Lista de ativos da empresa. }
  /users/me:
    get:
      tags: [Contexto]
      operationId: getCurrentUser
      security: [{ UserContext: [] }]
      responses:
        '200':
          description: Pessoa usuária autenticada (perfil/permissões).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }

  # ---- Ativos -----------------------------------------------------------
...
  /assets/{assetId}:
    patch:
      tags: [Ativos]
      operationId: updateAssetConfig
      description: Altera criticidade/configuração técnica. Ação de impacto.
      security: [{ UserContext: [] }]
      parameters:
        - name: assetId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ActionRequest'
                - type: object
                  properties:
                    changes:
                      type: object
                      properties:
                        criticality: { type: string, enum: [low, medium, high, critical] }
                        config: { $ref: '#/components/schemas/AssetConfig' }
      responses:
        '200':
          description: Configuração atualizada.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResult' }
        '400': { description: Justificativa ausente/inválida. }
        '403': { description: Sem permissão (action_high). }

  # ---- Análises ---------------------------------------------------------
  /assets/{assetId}/analyses:
    get:
      tags: [Análises]
      operationId: listAnalyses
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: assetId
          in: path
          required: true
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [current, stale, pending, inconclusive] }
      responses:
        '200': { description: Lista de análises do ativo. }
  /analyses/{analysisId}:
    get:
      tags: [Análises]
      operationId: getAnalysis
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: analysisId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Detalhe da análise (evidência, confiança, limitações).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryEnvelope' }
  /analyses/{analysisId}/reprocess:
    post:
      tags: [Análises]
      operationId: reprocessAnalysis
      description: Reprocessa análise. Ação de impacto; exige justificativa. Sucesso sem ciclo de status.
      security: [{ UserContext: [] }]
      parameters:
        - name: analysisId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionRequest' }
      responses:
        '200':
          description: Reprocesso aceito.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResult' }
        '400': { description: Justificativa ausente/inválida. }
  /analyses/{analysisId}/request-specialist:
    post:
      tags: [Análises]
      operationId: requestSpecialistAnalysis
      description: Solicita análise especializada. Ação de impacto; exige justificativa.
      security: [{ UserContext: [] }]
      parameters:
        - name: analysisId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionRequest' }
      responses:
        '200':
          description: Acionamento aceito.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResult' }

  # ---- Dados técnicos ---------------------------------------------------
  /assets/{assetId}/baseline:
    get:
      tags: [Dados técnicos]
      operationId: getBaseline
      description: |
        Estado "normal" aprendido do ativo/ponto. Permite explicar por que um insight não foi
        gerado (baseline em `learning`), por que está desatualizado (`invalidated` após manutenção)
        e como uma falha foi detectada (desvio vs. `symptom`). Falhas sintomáticas (lubrificação)
        retornam `detection_mode = symptom` e `learnable = false`.
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: assetId
          in: path
          required: true
          schema: { type: string }
        - name: point_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Baseline aprendido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryEnvelope' }
  /assets/{assetId}/rms:
    get:
      tags: [Dados técnicos]
      operationId: getRmsSeries
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: assetId
          in: path
          required: true
          schema: { type: string }
        - name: point_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Série temporal de RMS.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryEnvelope' }
  /assets/{assetId}/spectrum:
    get:
      tags: [Dados técnicos]
      operationId: getSpectrum
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: assetId
          in: path
          required: true
          schema: { type: string }
        - name: point_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Espectro FFT.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryEnvelope' }
  /assets/{assetId}/data-quality:
    get:
      tags: [Dados técnicos]
      operationId: getDataQuality
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: assetId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Qualidade/frescor dos dados.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryEnvelope' }

  # ---- Modelos ----------------------------------------------------------
  /models/{modelId}:
    get:
      tags: [Modelos]
      operationId: getModel
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: modelId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Versão, cobertura, requisitos, estado de processamento.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryEnvelope' }
  /models/{modelId}/request-retraining:
    post:
      tags: [Modelos]
      operationId: requestRetraining
      description: Solicita retreinamento. Ação de alto impacto; exige justificativa forte.
      security: [{ UserContext: [] }]
      parameters:
        - name: modelId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionRequest' }
      responses:
        '200':
          description: Solicitação aceita.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResult' }

  # ---- Conhecimento -----------------------------------------------------
  /knowledge/search:
    get:
      tags: [Conhecimento]
      operationId: searchKnowledge
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string, enum: [procedure, glossary, guidance] }
      responses:
        '200': { description: Documentos relevantes. }
  /knowledge/{docId}:
    get:
      tags: [Conhecimento]
      operationId: getKnowledgeDoc
      parameters:
        - $ref: '#/components/parameters/Seed'
        - name: docId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Documento de conhecimento. }

  # ---- Ações / Escalonamento -------------------------------------------
  /cases/{caseId}/escalate:
    post:
      tags: [Ações]
      operationId: escalateCase
      description: Encaminha o caso para análise humana. Ação de impacto; exige justificativa.
      security: [{ UserContext: [] }]
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionRequest' }
      responses:
        '200':
          description: Escalonamento aceito.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResult' }
security:
  - UserContext: []